pyproject.toml
[tool.poetry]
name = "online-training"
version = "0.1.0"
description = ""
authors = ["AlexandraAvtaykina <justalexandra@mail.ru>"]
readme = "README.md"
packages = [{include = "online_training"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
ipython = "^8.18.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.3.3"
drf-yasg = "^1.21.7"
stripe = "^7.10.0"
celery = "^5.3.6"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
